:recycle: Refactor demo to use Tailwind instead of Emotion
