+++ CONTRIBUTING.md
@@ -24,8 +24,7 @@ After cloning useWebAnimations, run `yarn` to fetch its dependencies. Then, you 
 
 - `yarn dev` runs the [demo app](https://github.com/wellyshen/use-web-animations/tree/master/demo/App) as your playground at `localhost:10001`. Support live reloading.
 - `yarn lint:code` lints all `.js` and `.tsx?` files.
-- `yarn lint:style` lints all `.css` and `.tsx?` (for [Emotion](https://emotion.sh)) files.
-- `yarn lint` lints both `code` and `style`.
+- `yarn lint` lints `code`.
 - `yarn type-check` runs the [Typescript](https://www.typescriptlang.org) typechecks.
 - `yarn type-check:watch` runs an interactive typechecks watcher (helpful in development).
 - `yarn test` runs the complete test suite.
