+++ demo/index.tsx
@@ -1,6 +1,6 @@
+import "./style.css";
 import React from "react";
-import ReactDOM from "react-dom";
-
+import { render } from "react-dom";
 import App from "./App";
 
-ReactDOM.render(<App />, document.getElementById("app"));
+render(<App />, document.getElementById("app"));
