+++ demo/style.css
@@ -1,0 +1,2 @@
+@tailwind base;
+@tailwind utilities;
