--- demo/utils/mq.ts
@@ -1,6 +1,0 @@
-export default {
-  sm: "@media (min-width: 576px)",
-  md: "@media (min-width: 768px)",
-  lg: "@media (min-width: 992px)",
-  xl: "@media (min-width: 1200px)",
-};
