+++ package.json
@@ -37,7 +37,6 @@
   "scripts": {
     "dev": "yarn clean:dev && rollup -c rollup/config.js -w --environment BUILD:dev",
     "lint:code": "eslint --fix . --ext .js,.ts,.tsx",
-    "lint:style": "stylelint --fix \"**/*.{css,ts,tsx}\"",
     "lint": "run-p lint:*",
     "type-check": "tsc --noEmit",
     "type-check:watch": "yarn type-check --watch",
@@ -63,7 +62,6 @@
   },
   "lint-staged": {
     "*.{js,ts,tsx}": "eslint --fix",
-    "*.{css,ts,tsx}": "stylelint --fix",
     "*.{json,md,yml}": "prettier --write"
   },
   "babel": {
@@ -75,16 +73,6 @@
     "plugins": [
       "@babel/proposal-object-rest-spread"
     ],
-    "env": {
-      "development": {
-        "presets": [
-          "@emotion/babel-preset-css-prop"
-        ],
-        "plugins": [
-          "emotion"
-        ]
-      }
-    },
     "ignore": [
       "src/polyfill.js"
     ]
@@ -102,8 +90,7 @@
             "error"
           ]
         }
-      ],
-      "emotion/jsx-import": "off"
+      ]
     }
   },
   "eslintIgnore": [
@@ -111,22 +98,6 @@
     "dist",
     "src/polyfill.js"
   ],
-  "stylelint": {
-    "plugins": [
-      "stylelint-prettier"
-    ],
-    "extends": [
-      "stylelint-config-standard",
-      "stylelint-config-prettier"
-    ],
-    "rules": {
-      "prettier/prettier": true
-    },
-    "ignoreFiles": [
-      "public/**/*.css",
-      "coverage/**/*.css"
-    ]
-  },
   "jest": {
     "preset": "ts-jest",
     "moduleNameMapper": {
@@ -143,20 +114,18 @@
     "@babel/preset-env": "^7.10.3",
     "@babel/preset-react": "^7.10.1",
     "@babel/preset-typescript": "^7.10.1",
-    "@emotion/babel-preset-css-prop": "^10.0.27",
-    "@emotion/core": "^10.0.28",
     "@rollup/plugin-babel": "^5.0.3",
     "@rollup/plugin-commonjs": "^13.0.0",
     "@rollup/plugin-html": "^0.2.0",
     "@rollup/plugin-node-resolve": "^8.0.1",
     "@rollup/plugin-replace": "^2.3.3",
     "@rollup/plugin-url": "^5.0.1",
+    "@tailwindcss/ui": "^0.3.0",
     "@testing-library/react-hooks": "^3.3.0",
     "@types/jest": "^26.0.0",
     "@types/react": "^16.9.38",
     "@types/react-dom": "^16.9.8",
     "@types/resize-observer-browser": "^0.1.3",
-    "babel-plugin-emotion": "^10.0.33",
     "eslint": "^6.8.0",
     "eslint-config-welly": "^1.2.0",
     "husky": "^4.2.5",
@@ -164,6 +133,7 @@
     "lint-staged": "^10.2.11",
     "normalize.css": "^8.0.1",
     "npm-run-all": "^4.1.5",
+    "postcss-discard-comments": "^4.0.2",
     "prettier": "^2.0.5",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
@@ -176,10 +146,7 @@
     "rollup-plugin-serve": "^1.0.1",
     "rollup-plugin-size-snapshot": "^0.12.0",
     "rollup-plugin-terser": "^6.1.0",
-    "stylelint": "^13.6.1",
-    "stylelint-config-prettier": "^8.0.2",
-    "stylelint-config-standard": "^20.0.0",
-    "stylelint-prettier": "^1.1.2",
+    "tailwindcss": "^1.4.6",
     "ts-jest": "^26.1.0",
     "typescript": "^3.9.5",
     "web-animations-js": "^2.3.2",
